[Refactor] combine test file
