+++ SSD_Driver/SSD_Driver.vcxproj.filters
@@ -30,10 +30,13 @@
     <ClCompile Include="fileio.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="ssd_read_test.cpp">
+    <ClCompile Include="ssd_test.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="ssd_test.cpp">
+    <ClCompile Include="bufferManager.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="bufferManager_test.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
   </ItemGroup>
@@ -48,6 +51,9 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="main.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="bufferManager.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
